No need to convert large polynomials to unsigned int.
Replaced all occurrences of uint64_t by unsigned int. No need to
force the CPU to use 64-bit words if it doesn't want to.

+++ ffs/f2pol.h
@@ -63,15 +63,15 @@
 /* Integer conversions.
  *****************************************************************************/
 
-// Conversion of an n-term polynomial to uint64_t.
+// Conversion of an n-term polynomial to an unsigned int.
 #define __FP_GET_UI(sz, r, p, n) \
-  do { r = (uint64_t)p[0]; } while (0)
+  do { r = (unsigned)p[0]; } while (0)
 
-// Conversion of an n-term polynomial from uint64_t.
+// Conversion of an n-term polynomial from an unsigned int.
 #define __FP_SET_UI(sz, r, x, n) \
   do { r[0] = (uint##sz##_t)x; } while (0)
 
-// Conversions to/from uint64_t in the case of monic polynomials.
+// Conversions to/from an unsigned int in the case of monic polynomials.
 #define __FP_MONIC_GET_UI __FP_GET_UI
 #define __FP_MONIC_SET_UI __FP_SET_UI
